refactoring server.py
