+++ backend/geonature/__init__.py
@@ -8,3 +8,6 @@ conf = load_config(DEFAULT_CONFIG_FILE)
 root_logger = logging.getLogger()
 root_logger.addHandler(logging.StreamHandler())
 root_logger.setLevel(conf["SERVER"]["LOG_LEVEL"])
+
+
+from .server import create_app
