+++ backend/geonature/utils/utilsmails.py
@@ -3,7 +3,7 @@
 from flask import current_app
 from flask_mail import Message
 
-from server import MAIL
+from geonature.server import MAIL
 
 
 def send_mail(recipients, subject, msg_html):
