+++ backend/tests/bootstrap_test.py
@@ -5,7 +5,7 @@ import pytest
 from flask import url_for, current_app
 from cookies import Cookie
 
-import server
+from geonature import create_app
 from geonature.utils.env import load_config, get_config_file_path
 
 # TODO: fixture pour mettre des données test dans la base a chaque test
@@ -15,7 +15,7 @@ from geonature.utils.env import load_config, get_config_file_path
 def app():
     config_path = get_config_file_path()
     config = load_config(config_path)
-    app = server.get_app(config)
+    app = create_app(config)
     app.config["TESTING"] = True
     app.config["WTF_CSRF_ENABLED"] = False
     return app
