+++ backend/tests/conftest.py
@@ -3,14 +3,14 @@ import os, logging
 import psycopg2
 
 from geonature.utils.env import load_config, get_config_file_path
-import server
+from geonature import create_app
 
 
 def pytest_sessionstart(session):
     """ before session.main() is called. """
     config_path = get_config_file_path()
     config = load_config(config_path)
-    app = server.get_app(config)
+    app = create_app(config)
     app.config["TESTING"] = True
     # push the app_context
     ctx = app.app_context()
